perfetto: Address review comments on IRQ mapping data source.
- Use StringSplitter instead of manual char iteration
- Pass /proc/interrupts content via ReadFile() for testability
- Extend trigger detection to cover x86 (-edge, -fasteoi) formats
- Switch irq_id to uint32, use FlatHashMap, remove dead update path
- Add unit tests (ARM + x86) and stdlib diff test

+++ src/trace_processor/importers/proto/system_probes_parser.cc
@@ -1205,8 +1205,7 @@ void SystemProbesParser::ParseInterruptInfo(ConstBytes blob) {
     protos::pbzero::InterruptInfo::InterruptMapping::Decoder mapping(*it);
     if (mapping.has_irq_id() && mapping.has_name()) {
       auto name_id = context_->storage->InternString(mapping.name());
-      context_->irq_tracker->SetIrqName(static_cast<uint32_t>(mapping.irq_id()),
-                                        name_id);
+      context_->irq_tracker->SetIrqName(mapping.irq_id(), name_id);
     }
   }
 }
